+++ src/main/java/org/folio/service/PatronBlocksService.java
@@ -1,17 +1,17 @@
 package org.folio.service;
 
+import static io.vertx.core.Future.failedFuture;
 import static io.vertx.core.Future.succeededFuture;
 import static java.lang.String.format;
-import static org.folio.domain.Condition.isConditionLimitExceeded;
 import static org.folio.okapi.common.XOkapiHeaders.TENANT;
 
 import java.lang.invoke.MethodHandles;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
-import org.folio.exception.EntityNotFoundException;
+import org.folio.domain.ActionBlocks;
 import org.folio.repository.PatronBlockConditionsRepository;
 import org.folio.repository.PatronBlockLimitsRepository;
 import org.folio.repository.UserSummaryRepository;
@@ -24,14 +24,17 @@ import org.folio.rest.jaxrs.model.UserSummary;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.tools.utils.TenantTool;
 
-import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+import lombok.With;
 
 public class PatronBlocksService {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+  private static final String DEFAULT_ERROR_MESSAGE = "Failed to calculate automated patron blocks";
 
   private final UserSummaryRepository userSummaryRepository;
   private final PatronBlockConditionsRepository conditionsRepository;
@@ -50,59 +53,155 @@ public class PatronBlocksService {
   public Future<AutomatedPatronBlocks> getBlocksForUser(String userId) {
     return userSummaryRepository.getByUserId(userId)
       .compose(optionalSummary -> optionalSummary
+        .map(userSummary -> new BlocksCalculationContext().withUserSummary(userSummary))
         .map(this::getBlocksForSummary)
         .orElseGet(() -> succeededFuture(new AutomatedPatronBlocks())));
   }
 
-  private Future<AutomatedPatronBlocks> getBlocksForSummary(UserSummary summary) {
-    return succeededFuture(summary.getUserId())
-      .compose(usersClient::findPatronGroupIdForUser)
-      .compose(limitsRepository::findLimitsForPatronGroup)
-      .compose(limits -> calculateBlocks(summary, limits));
+  private Future<AutomatedPatronBlocks> getBlocksForSummary(BlocksCalculationContext ctx) {
+    return succeededFuture(ctx)
+      .compose(this::addUserGroupIdToContext)
+      .compose(this::addPatronBlockLimitsToContext)
+      .compose(this::addAllPatronBlockConditionsToContext)
+      .map(this::calculateBlocks);
   }
 
-  private Future<AutomatedPatronBlocks> calculateBlocks(UserSummary summary,
-    List<PatronBlockLimit> limits) {
-
+  private AutomatedPatronBlocks calculateBlocks(BlocksCalculationContext ctx) {
     final AutomatedPatronBlocks blocks = new AutomatedPatronBlocks();
 
-    if (limits.isEmpty()) {
-      return succeededFuture(blocks);
+    if (ctx.patronBlockLimits.isEmpty()) {
+      return blocks;
     }
 
-    List<Future<AutomatedPatronBlock>> futures = limits.stream()
-      .filter(limit -> isConditionLimitExceeded(summary, limit))
+    blocks.getAutomatedPatronBlocks().addAll(ctx.patronBlockLimits.stream()
+      .map(ctx::withCurrentPatronBlockLimit)
+      .map(this::addCurrentConditionToContext)
+      .map(this::addActionBlocksByLimitAndConditionToContext)
+      .filter(context -> context.currentActionBlocks.isNotEmpty())
       .map(this::createBlockForLimit)
-      .collect(Collectors.toList());
+      .filter(Objects::nonNull)
+      .collect(Collectors.toList()));
 
-    return CompositeFuture.all(new ArrayList<>(futures))
-      .onFailure(log::error)
-      .onSuccess(cf -> blocks.getAutomatedPatronBlocks().addAll(cf.list()))
-      .map(blocks);
+    return blocks;
   }
 
-  private Future<AutomatedPatronBlock> createBlockForLimit(PatronBlockLimit limit) {
-    return succeededFuture(limit)
-      .compose(this::findConditionForLimit)
-      .map(this::createBlockForCondition);
+  private Future<BlocksCalculationContext> addUserGroupIdToContext(BlocksCalculationContext ctx) {
+    if (ctx.userSummary == null || ctx.userSummary.getUserId() == null) {
+      logAddingToContextError("user group ID");
+      return failedFuture(DEFAULT_ERROR_MESSAGE);
+    }
+
+    return usersClient.findPatronGroupIdForUser(ctx.userSummary.getUserId())
+      .map(ctx::withUserGroupId);
   }
 
-  private Future<PatronBlockCondition> findConditionForLimit(PatronBlockLimit limit) {
-    final String conditionId = limit.getConditionId();
+  private Future<BlocksCalculationContext> addPatronBlockLimitsToContext(
+    BlocksCalculationContext ctx) {
 
-    return conditionsRepository.get(conditionId)
-      .map(optional -> optional.orElseThrow(() ->
-        new EntityNotFoundException(format(
-          "Condition %s referenced by limit %s does not exist", conditionId, limit.getId()))));
+    if (ctx.userGroupId == null) {
+      logAddingToContextError("patron block limits");
+      return failedFuture(DEFAULT_ERROR_MESSAGE);
+    }
+
+    return limitsRepository.findLimitsForPatronGroup(ctx.userGroupId)
+      .map(ctx::withPatronBlockLimits);
   }
 
-  private AutomatedPatronBlock createBlockForCondition(PatronBlockCondition condition) {
+  private Future<BlocksCalculationContext> addAllPatronBlockConditionsToContext(
+    BlocksCalculationContext ctx) {
+
+    return conditionsRepository.getAllWithDefaultLimit().map(ctx::withPatronBlockConditions);
+  }
+
+  private BlocksCalculationContext addCurrentConditionToContext(
+    BlocksCalculationContext ctx) {
+
+    if (ctx.currentPatronBlockLimit == null ||
+      ctx.currentPatronBlockLimit.getConditionId() == null) {
+
+      logAddingToContextError("current condition");
+      return ctx;
+    }
+
+    String conditionId = ctx.currentPatronBlockLimit.getConditionId();
+
+    PatronBlockCondition patronBlockCondition = ctx.patronBlockConditions.stream()
+      .filter(condition -> condition.getId().equals(conditionId))
+      .findFirst()
+      .orElse(null);
+
+    if (patronBlockCondition == null) {
+      logAddingToContextError("current condition",
+        format("cannot find condition by ID %s", conditionId));
+      return ctx;
+    }
+
+    return ctx.withCurrentPatronBlockCondition(patronBlockCondition);
+  }
+
+  private BlocksCalculationContext addActionBlocksByLimitAndConditionToContext(
+    BlocksCalculationContext ctx) {
+
+    if (ctx.userSummary == null || ctx.currentPatronBlockLimit == null ||
+      ctx.currentPatronBlockCondition == null) {
+
+      logAddingToContextError("action blocks by limit and condition");
+      return ctx;
+    }
+
+    PatronBlockCondition patronBlockCondition = ctx.currentPatronBlockCondition;
+
+    ActionBlocks actionBlocksByLimit = ActionBlocks.byLimit(ctx.userSummary,
+      ctx.currentPatronBlockLimit);
+
+    ActionBlocks actionBlocksByCondition = new ActionBlocks(
+      Boolean.TRUE.equals(patronBlockCondition.getBlockBorrowing()),
+      Boolean.TRUE.equals(patronBlockCondition.getBlockRenewals()),
+      Boolean.TRUE.equals(patronBlockCondition.getBlockRequests()));
+
+    return ctx.withCurrentActionBlocks(
+      ActionBlocks.and(actionBlocksByLimit, actionBlocksByCondition));
+  }
+
+  private AutomatedPatronBlock createBlockForLimit(BlocksCalculationContext ctx) {
+    if (ctx.currentPatronBlockCondition == null || ctx.currentActionBlocks == null) {
+      log.error("Failed to create automated patron block - context is invalid");
+      return null;
+    }
+
     return new AutomatedPatronBlock()
-      .withPatronBlockConditionId(condition.getId())
-      .withBlockBorrowing(condition.getBlockBorrowing())
-      .withBlockRenewals(condition.getBlockRenewals())
-      .withBlockRequests(condition.getBlockRequests())
-      .withMessage(condition.getMessage());
+      .withPatronBlockConditionId(ctx.currentPatronBlockCondition.getId())
+      .withBlockBorrowing(ctx.currentActionBlocks.getBlockBorrowing())
+      .withBlockRenewals(ctx.currentActionBlocks.getBlockRenewals())
+      .withBlockRequests(ctx.currentActionBlocks.getBlockRequests())
+      .withMessage(ctx.currentPatronBlockCondition.getMessage());
   }
 
+  private void logAddingToContextError(String parameterName, String additionalMessage) {
+    String message = format("Blocks calculation context is invalid, failed to add %s",
+      parameterName);
+
+    if (additionalMessage != null) {
+      message += format(". %s", additionalMessage);
+    }
+
+    log.error(message);
+  }
+
+  private void logAddingToContextError(String parameterName) {
+    logAddingToContextError(parameterName, null);
+  }
+
+  @With
+  @AllArgsConstructor
+  @NoArgsConstructor(force = true)
+  private static class BlocksCalculationContext {
+    final UserSummary userSummary;
+    final String userGroupId;
+    final List<PatronBlockLimit> patronBlockLimits;
+    final List<PatronBlockCondition> patronBlockConditions;
+    final PatronBlockLimit currentPatronBlockLimit;
+    final PatronBlockCondition currentPatronBlockCondition;
+    final ActionBlocks currentActionBlocks;
+  }
 }